Added Distortion Shader, Basic Beatmapper
Brought in the distortion shader which must be set as the type of a material. Can be used with the DistortObject script.
Brought in a basic 'beat mapper' that responds to the bass. Works really well for the first song, and less well for the others. Will need refinement.

Known Shader Issues:
Currently distorts in local space, not screen space.
Distorion effect doesn't work properly on simple models with few verticies (or rather can't make a wave).
Shader attempts to divide by zero (despite being told not to), causing brief but large errors during pulses.
Will need to be merged with existing shaders for certain objects (pyramids, weapons).
Pulsing seems to ignore BeatsBetweenPulse, though that's probably an issue with the beat mapper.